requirements.txt
numpy==1.19.0
pandas==1.0.5
Pillow==7.1.2
scipy==1.4.1
streamlit==0.61.0
tensorflow==2.2.0